Move Memory files to shared and add Arena/MemorySegment stubs
